requirements.txt
torch>=2.0
numpy
pandas
tqdm
fair-esm
